WIP: Assimp loads files, but few things need to be sorted out
In example 004 the code works but:

Model file is not in yet, becuase I am not too sure if
Blender file will do, or does it necessary have to be
exported to something else.

The reason of this is that normals are not loaded from file
and I am finding way to have them.

Assimp can generate normals, but It has some glithes, I seek to  make it possible to load it from model file.

Also, possibly mapToWorld is not regarded in light position.